Now module loading system implemented
