+++ modules/maze.py
@@ -191,3 +191,5 @@ class Maze(modules.Module):
     COMMANDS = {
         "move": cmd_move,
     }
+
+__module_class__=Maze
